plugins/gauntlet/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gauntlet-dev"
version = "1.8.1"
description = "Codebase learning through knowledge extraction, challenges, and spaced repetition"
authors = [
    {name = "Alex Thola", email = "alexthola@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pyyaml>=6.0",
    "tree-sitter>=0.23,<1",
    "tree-sitter-language-pack>=0.3,<1",
    "networkx>=3.2,<4",
]

[project.optional-dependencies]
graph = [
    "igraph>=0.11,<1",
]
dev = [
    "ruff>=0.14.13",
    "mypy>=1.11.0",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "types-pyyaml>=6.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/gauntlet"]

[tool.ruff]
line-length = 88
target-version = "py39"
exclude = [".venv", ".uv-cache"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "C4", "UP", "S", "PL", "D"]
extend-ignore = ["S101", "E203", "D203", "D213", "UP017"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "S101", "S108", "PLR2004", "PLC0415", "D102", "D103", "D205", "D400",
    "D415", "D212", "S603", "S607", "PLW1510",
]
"scripts/**/*.py" = ["S603", "S607", "PLR0911", "PLR2004", "UP045", "S105"]
"hooks/**/*.py" = ["S603", "S607", "PLR0911", "PLR2004", "UP045", "S105", "PLC0415", "BLE001", "E402"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
testpaths = ["tests"]
pythonpath = ["src", "scripts", "hooks"]
addopts = "-v --cov=src/gauntlet --cov-report=term-missing --cov-fail-under=90"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
]

[tool.coverage.run]
source = ["src/gauntlet"]
omit = ["tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
fail_under = 90

[dependency-groups]
dev = [
    "mypy>=1.11.0",
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.14.13",
    "types-pyyaml>=6.0.12.20250915",
]
